Created Storage File and Broke Node into Files
Created an abstract collection called Storage and ArrayStorage which
inherits it and I broke Node class into multiple files for better
readability.
